net: dsa: sja1105: Simplify reset handling
We don't really need 10k species of reset. Remove everything except cold
reset which is what is actually used. Too bad the hardware designers
couldn't agree to use the same bit field for rev 1 and rev 2, so the
(*reset_cmd) function pointer is there to stay. However let's simplify
the prototype and give it a struct dsa_switch (we don't want
forward-declarations of struct sja1105_private).

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>