net: dsa: sja1105: Request SPI timestamp for correct word
With the change to use bits_per_word=32, the sja1105 driver no longer
uses the correct API for PTP system timestamping (something that is
required for proper cross-timestamping with the value of the system
clock at the time of the SPI transactions).

This fixes a variety of random PTP bugs, most notably clock stepping
with phc2sys if SYSOFF_EXTENDED is being used.

Fixes: 23ec4bb3ab87 ("net: dsa: sja1105: Increase SPI word size to 32 bits")
Signed-off-by: Vladimir Oltean <olteanv@gmail.com>