arm64: dts: fsl: ls1028a-rdb: Free up eno2/swp4 for 802.1CB by enabling eno3/swp5
This patch moves the CPU port from swp4 to swp5 on the Felix switch, and
uses the 2.5 port pair which used to be for DSA tagging in plain
untagged mode.

The switch core for 802.1CB FRER uses the MAC table (FDB) to assign
packets to a Seamless Stream ID (SSID). In turn, the MAC table needs the
Analyzer module (ANA) to inspect the frames.

But traffic that passes to/from the DSA CPU port has the BYPASS flag set
in the injection header, which means "bypass the analyzer module".

So traffic originated from the CPU will not be correctly assigned to a
SSID for 802.1CB if it is coming from the port pair where DSA tagging is
enabled.

As a workaround, we enable the other port pair and use that for 802.1CB
traffic that needs to be terminated on the local CPU.

Before:                               After:

    eno2    eno3                         eno2      eno3
     |       |                   Untagged |         |
     | DSA   x disabled           traffic |         | DSA
     | tags  |                  including |         | tags
     |       |                    802.1CB |         |
    swp4    swp5                         swp4      swp5

Of course, in this context "untagged" means "not tagged with an
injection/extraction DSA tag", it has nothing to do with the VLAN or
802.1CB R-Tag.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>